refactor(WinningLotto): WinningLotto 의 패키지를 domain 으로 변경
WinningLotto 는 서비스 역할이 아닌 도메인 역할에 더 맞다고 판단해 패키지를 변경함.
